+++ src/comm/publish/implement.rs
@@ -1,12 +1,18 @@
-use std::{net::IpAddr, vec};
+use std::net::IpAddr;
 
 use anyhow::{Result, anyhow, bail};
 use giganto_client::publish::stream::{
-    RequestSemiSupervisedStream, RequestTimeSeriesGeneratorStream,
+    RequestSemiSupervisedStream, RequestStreamRecord, RequestTimeSeriesGeneratorStream,
 };
 
+use crate::comm::stream_channel_key::StreamChannelKey;
+
 pub trait RequestStreamMessage {
-    fn channel_key(&self, sensor: Option<String>, record_type: &str) -> Result<Vec<String>>;
+    fn channel_keys(
+        &self,
+        publisher_sensor: Option<&str>,
+        record_type: RequestStreamRecord,
+    ) -> Result<Vec<StreamChannelKey>>;
     fn start_time(&self) -> i64;
     fn filter_ip(&self, orig_addr: IpAddr, resp_addr: IpAddr) -> bool;
     fn sensor(&self) -> Result<String>;
@@ -16,25 +22,23 @@ pub trait RequestStreamMessage {
 }
 
 impl RequestStreamMessage for RequestSemiSupervisedStream {
-    fn channel_key(&self, sensor: Option<String>, record_type: &str) -> Result<Vec<String>> {
-        let sensor = sensor.ok_or_else(|| {
+    fn channel_keys(
+        &self,
+        publisher_sensor: Option<&str>,
+        record_type: RequestStreamRecord,
+    ) -> Result<Vec<StreamChannelKey>> {
+        let publisher_sensor = publisher_sensor.ok_or_else(|| {
             anyhow!("Failed to generate the Semi-supervised channel key, sensor is required.")
         })?;
         if let Some(ref sensor_list) = self.sensor {
             let semi_supervised_keys = sensor_list
                 .iter()
-                .map(|target_sensor| {
-                    let mut key = String::new();
-                    key.push_str("SemiSupervised");
-                    key.push('\0');
-                    key.push_str(&sensor);
-                    key.push('\0');
-                    key.push_str(target_sensor);
-                    key.push('\0');
-                    key.push_str(record_type);
-                    key
+                .map(|target_sensor| StreamChannelKey::SemiSupervised {
+                    publisher_sensor: publisher_sensor.to_string(),
+                    target_sensor: target_sensor.clone(),
+                    record_type,
                 })
-                .collect::<Vec<String>>();
+                .collect::<Vec<StreamChannelKey>>();
             return Ok(semi_supervised_keys);
         }
         bail!("Failed to generate the Semi-supervised Engine channel key, sensor is required.");
@@ -68,17 +72,17 @@ impl RequestStreamMessage for RequestSemiSupervisedStream {
 }
 
 impl RequestStreamMessage for RequestTimeSeriesGeneratorStream {
-    fn channel_key(&self, _sensor: Option<String>, record_type: &str) -> Result<Vec<String>> {
+    fn channel_keys(
+        &self,
+        _publisher_sensor: Option<&str>,
+        record_type: RequestStreamRecord,
+    ) -> Result<Vec<StreamChannelKey>> {
         if let Some(ref target_sensor) = self.sensor {
-            let mut time_series_generator_key = String::new();
-            time_series_generator_key.push_str("TimeSeriesGenerator");
-            time_series_generator_key.push('\0');
-            time_series_generator_key.push_str(&self.id);
-            time_series_generator_key.push('\0');
-            time_series_generator_key.push_str(target_sensor);
-            time_series_generator_key.push('\0');
-            time_series_generator_key.push_str(record_type);
-            return Ok(vec![time_series_generator_key]);
+            return Ok(vec![StreamChannelKey::TimeSeriesGenerator {
+                id: self.id.clone(),
+                target_sensor: target_sensor.clone(),
+                record_type,
+            }]);
         }
         bail!("Failed to generate the Time Series Generator channel key, sensor is required.");
     }